Moved param-related classes into `TLAW::Params` namespace
