+++ lib/tlaw/params/keyword.rb
@@ -1,0 +1,19 @@
+require_relative 'param'
+
+module TLAW
+  module Params
+    class Keyword < Param
+      def keyword?
+        true
+      end
+
+      def to_code
+        if required?
+          "#{name}:"
+        else
+          "#{name}: #{default_to_code}"
+        end
+      end
+    end
+  end
+end
